backend/requirements.txt
# Build tools (required for some packages)
setuptools>=75.0.0
wheel>=0.44.0

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# News collection and processing
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
requests-cache==1.1.1
lxml==4.9.3

# Data processing and analysis
numpy==1.26.4
pandas==2.0.3
python-dateutil==2.8.2

# NLP and keyword extraction (optional)
kiwipiepy==0.16.2

# Wordcloud generation
wordcloud==1.9.3
pillow==10.1.0
matplotlib==3.7.4

# AI/ML services (optional)
openai==1.3.0

# Utility packages
pathlib2==2.3.7
urllib3>=1.26.0,<2.0.0

# Production server
gunicorn==21.2.0
asgiref==3.7.2